Small refactor.  Shortened the name of some of the collections files to take advantage of the namspace for context
