+++ ragchat/application/use_cases/collections/use_cases.py
@@ -1,5 +1,5 @@
 from ragchat.common.exceptions.entity_not_found_exception import EntityNotFoundException
-from ragchat.domain.collections.collection_repository_interface import ICollectionRepository
+from ragchat.domain.collections.repository_interface import ICollectionRepository
 from ragchat.domain.collections.collection import Collection
 from .dtos import ListCollectionsViewModel
 from uuid import UUID, uuid4
